revert allat

+++ private/api/gameservers/close.php
@@ -1,17 +1,16 @@
 <?php
 	use anorrl\GameServer;
 
-	use anorrl\utilities\ClientDetector;
+	$access = CONFIG->asset->key;
 
-	if(!ClientDetector::HasAccess())
-		exit(http_response_code(500));
+	if(isset($_GET['access']) && isset($_GET['jobID'])) {
+		if($_GET['access'] == $access) {
+			$gameserver = GameServer::GetFromJobID($_GET['jobID']);
 
-	if(isset($_GET['jobID'])) {
-		$gameserver = GameServer::GetFromJobID($_GET['jobID']);
-
-		if($gameserver) {
-			$gameserver->destroy();
-			die();
+			if($gameserver) {
+				$gameserver->destroy();
+				die();
+			}
 		}
 	}
